Only use local WSL Bash for Automake and Autoconf
Automake and Autoconf are still giving me a lot of problems
when using them on Windows. I seem to get better results
when I force Conan to use the local WSL Bash, which is not
ideal, but this unblocks me for now.

See also conan-io/conan#9222 , conan-io/conan#10028 and
conan-io/conan#9048

Contributes to CURA-8831